refactor: [#100] rename module to media_library_filename.py
