+++ src/nautilus_librarian/domain/dvc_services_api.py
@@ -2,7 +2,9 @@ from os import path
 from typing import List
 
 from nautilus_librarian.mods.dvc.domain.api import DvcApiWrapper
-from nautilus_librarian.mods.namecodes.domain.filename import MediaLibraryFilename
+from nautilus_librarian.mods.namecodes.domain.media_library_filename import (
+    MediaLibraryFilename,
+)
 from nautilus_librarian.mods.namecodes.domain.validate_filenames import (
     is_a_library_file,
 )
