+++ src/nautilus_librarian/domain/file_locator.py
@@ -1,6 +1,8 @@
 import os
 
-from nautilus_librarian.mods.namecodes.domain.filename import MediaLibraryFilename
+from nautilus_librarian.mods.namecodes.domain.media_library_filename import (
+    MediaLibraryFilename,
+)
 
 
 class FileNotFoundException(Exception):
