+++ src/nautilus_librarian/domain/validate_filepaths.py
@@ -1,7 +1,9 @@
 import os
 
 from nautilus_librarian.domain.file_locator import file_locator
-from nautilus_librarian.mods.namecodes.domain.filename import MediaLibraryFilename
+from nautilus_librarian.mods.namecodes.domain.media_library_filename import (
+    MediaLibraryFilename,
+)
 
 
 class InvalidImageFolderException(ValueError):
