+++ src/nautilus_librarian/typer/commands/workflows/actions/rename_base_images_action.py
@@ -8,7 +8,9 @@ from nautilus_librarian.domain.file_locator import (
     get_base_image_absolute_path_from_gold,
     guard_that_base_image_exists,
 )
-from nautilus_librarian.mods.namecodes.domain.filename import MediaLibraryFilename
+from nautilus_librarian.mods.namecodes.domain.media_library_filename import (
+    MediaLibraryFilename,
+)
 from nautilus_librarian.typer.commands.workflows.actions.action_result import (
     ActionResult,
     ErrorMessage,
