+++ tests/test_nautilus_librarian/test_mods/test_namecodes/test_domain/test_media_library_filename.py
@@ -1,4 +1,6 @@
-from nautilus_librarian.mods.namecodes.domain.filename import MediaLibraryFilename
+from nautilus_librarian.mods.namecodes.domain.media_library_filename import (
+    MediaLibraryFilename,
+)
 
 
 def test_instantiation_from_string():
